0x02

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -11,9 +11,9 @@
 int main(void)
 {
 	int count;
-	long long int sum = 2;
-	long long int x = 1;
-	long long int c;
+	long int sum = 2;
+	long int x = 1;
+	long int c;
 
 	printf("1");
 	for (count = 1; count <= 50; count++)
